2.1 Hello World with Thor
The first framework that we.ll cover is Thor. It is a framework that builds command-line applications with ease. It is used by many other applications, one of them being Ruby on Rails.

In this lesson, we.ll dip our feet in the water by building a rather simple Hello World example.