3.1 - Introduction to GLI
We generate this with gem: GLI; in which it istalled: gem install gli

After the gem is installed, we can generate the files in this commit by

```
$ gli init recipes add list
```

This will create a recipes/bin/recipes file that contains switch like:

desc
```switch [:s, switch]``` in Thor this would writen as ```option :s, type: :boolean```

Aliases are built into the GLI with the use of arrays.

Flags are the same as using ```option``` in Thor wthout using the boolean type.

But GLI has pre, post and on_error methods that are ran before, after or whenever an error occur; repectively.

In recipes/lib/recipes.rb, is where the bulk of the commands domain logic resides where recipes/bin/recipes is the interface to that domain. This file just describes the commands at high level and separates interfacing from the real applicatin in ./lib.